Report failures properly for `oracle_user`
A lot of the code in `oracle_user` was passing around `msg` as if
its contents could be modified by assignment (they can't be).  I
removed all of the faulty `msg`-passing and made critical functions
fail properly if necessary.

The net result is that, on failure, there will always be a proper
`msg` provided.  Previously, this would often result in an empty
message, which is very difficult to troubleshoot.